modules/wb_acq_core: fix multishot triggered acquisition when aligning
For multishot triggered acquisitons we can't and
don't need to align the trigger to match the
samples atom requirements. This is due to the fact
that when using multishot, internal BRAM is used
and this requirement is satisfied by design.

In this case, we transfer the exact amount
of samples to acq_fc_fifo. So, we can't try to
correct for the alignment offset calculated by
acq_trigger as there is nothing to be corrected.
If we do try to correct we will end up reading
invalid samples.