wb_acq_core: fix requirement of valid sample in POST_TRIG state for multishot RAM
This is related to a2b586914f. Even with that fix, we still
required an additional valid sample for asserting acq_post_trig_done
signal in acq_fsm module. As this condition is only used by
acq_multishot_dpram (and not acq_fc_fifo module, used when DPRAM
is bypassed) to signal the end of the acquisition and start of the
DPRAM read, an additional valid sample even was needed even if
all samples were already acquired.