common/mov_avg_dyn: add core
A moving average filter with dynamic order setting.

This core only supports orders which the number of taps of
resulting FIR (order + 1) is a power of 2. The reason for that is so we
can perfom divisions by simply shifting the fixed point to the left.

Special thanks to @GSRamirez95 for generating the test files.