Implement a wishbone master interface controlled via UART
This interface is way simpler than wb_rs232_syscon, and its protocol
is computer friendly, so writing a parser to it should be trivial.

The main motivation for this is the fact that wb_rs232_syscon seems to
hang indefinitely under heavy usage, and its text protocol is not easy
to parse.